make row selectable
